Define IDebugProfileLaunchTargetsProvider4
Currently, the `LaunchProfilesDebugLaunchProvider` calls `IDebugProfileLaunchTargetsProvider.OnBeforeLaunchAsync` right before launching processes, and `IDebugProfileLaunchTargetsProvider.OnAfterLaunchAsync` right after. The idea is that, in addition to creating the launch settings for the processes, the launch targets provider may want to do something special right before or right after processes launch, like turn on some logging or show some UI.

There is some information available that isn't currently passed to the launch targets provider. One is the set of `IDebugLaunchSettings` objects that is actually going to be passed to CPS (and then on to the debugger) to start the processes, and the other is the set of `VsDebugTargetProcessInfo`s providing info about the started processes. We're going to need these, especially the latter, for Hot Reload scenarios. For example, when we start a process with Hot Reload enabled we need to know when it exits so that we can end the Hot Reload session. To do this, we need the process ID from the `VsDebugTargetProcessInfo`.

This change introduces `IDebugProfileLaunchTargetsProvider4` which provides replacements for `IDebugProfileLaunchTargetsProvider.OnBeforeLaunchAsync` and `IDebugProfileLaunchTargetsProvider.OnAfterLaunchAsync` that pass along the relevant information.