Implement IDebugProfileLaunchTargetsProvider4
Update `ProjectLaunchTargetsProvider` to implement `IDebugProfileLaunchTargetsProvider4`. Right now the type doesn't do anything interesting in the original versions of `OnBeforeLaunchAsync` and `OnAfterLaunchAsync`, and neither does it do anything interesting in the new versions. However, the original versions have been updated to throw an exception, as we don't expect them to be called.