Fast up-to-date check works on solution open
Previously, the first up-to-date check that runs after a project loads always reports that a build is required. This avoided an uncommon yet serious potential issue.

Consider a project that brings in all `.cs` files via a glob, with the following steps:

1. Load the project
1. Build the project
1. Close the project
1. Delete one of the input `.cs` files from disk
1. Reopen the project

If the fast up-to-date check only considered the timestamps of input and output files, the project would appear up-to-date. However, it is not.

If an item is removed while Visual Studio has the project loaded, the change is observed and handled correctly.

Note that it is also possible to trigger this situation by adding a file, however usually when a file is added its timestamp is updated. Of course software can set the timestamp to whatever it wants. One easy way to do this is to undelete a file, which restores it with the original timestamp.

In order to correctly handle the above scenario, we need to persist some information about the project's inputs, so that when it is reloaded we can decide whether any inputs have been removed from disk.

That is what this change does. We introduce `IUpToDateCheckStatePersistence` which stores state in the `.vs` folder. That state contains a hash of the items in each project configuration.

With this change, we are now able to correctly identify whether a reloaded project is up-to-date or not, and avoid the precautionary build.

For OrchardCore, this cuts the time for first build after solution load from just over a minute to around one second, in the case that all 157 projects were already up-to-date.

This change will also reduce wait times when running unit tests after loading a solution, or launching a debugger. It will also reduce wait times when building after the solution being reloaded, perhaps due to the `.sln` file changing during a branch switch.