Make LaunchProfileProjectProperties Rule-aware
An implementation of `IProjectProperties` may also optionally implement `IRuleAwareProjectProperties` in which case it will be provided a `Rule` instance that can provide additional information on how to interpret and store property values. I had always intended to implement this interface (and implementations of `IGlobalSettingExtensionValueProvider` and `ILaunchProfileExtensionValueProvider` already expect to receive the `Rule` if one is available) but never got around to it. Now that we want to read and write basic properties to `ILaunchProfile.OtherSettings` without requiring an implementation of `ILaunchProfileExtensionValueProvider` it is time.

At the `IProjectProperties` level property values are always handled as strings. Part of what implementations of `IGlobalSettingExtensionValueProvider` and `ILaunchProfileExtensionValueProvider` do is convert those strings to whatever type is appropriate to store in the `ILaunchProfile` and/or `ILaunchSettings`. If we're going to add a path that does not depend on those providers, then we need some other way of determinining the storage type of the property value. Conveniently, the property type is already encoded in the `Rule`--we just need access to it.

This also includes unit tests to verify that we do, in fact, pass the `Rule` to the extension providers.