Utilize the Rule when reading values
When reading a property from the `LaunchProfilePropertiesProvider` and it is not handled by either a launch profile value provider _or_ a global setting provider we want to try and read it from `ILaunchProfile.OtherSettings`. This is an `ImmutableDictionary<string, object>` and so we need to know how to convert the `object` to the `string` representation returned by the `LaunchProfilePropertiesProvider`. To do this, we make use of the `Rule` associated with the provider, if any.

Basically, we try to find a property of the same name in the `Rule`. If we do we switch on the type of the property to determine how to translate the value into the final string. If no `Rule` has been given or we can't find a matching property in the `Rule` we default to treating the value as a `string`. If conversion to the expected type fails we simply return `null`.

There are a couple of improvements we could make here:

1. We don't currently support the `StringListProperty` type. Right now we don't use it anywhere, but support could certainly be added in the future.
2. We don't make use of default values in the `Rule`. Again, this could be done but right now we don't have a need.