Send telemetry on a threadpool thread
During shutdown we compute and publish telemetry about project dependencies. Previously this code would run on the UI thread, although that is not necessary.

This change switches to a background thread before doing that work, freeing the UI thread up to do other work. It still blocks the caller until that work completes.