Fast up-to-date check uses latest data
We use Dataflow blocks to process project data through various transformations, ultimately producing the data needed for the fast up-to-date check. Dataflow processes data asynchronously, fanned across threads.

Previously the output of the Dataflow "graph" was cached to a field on the subscription type. When a query was made, this field was read. It was possible that the Dataflow graph hadn't finished pushing data through its transformations at the time the query was made, meaning the query's result could be based on old data.

With this change, the fast up-to-date check now ensures that the most recent configured project data has flowed through the graph, so that the answer is always based on the latest data.

One complexity of this change relates to Visual Studio's "command line" mode, where `devenv.exe` could be used to perform builds. In such a case, the up-to-date check is called. However, design-time builds are not available in command line mode, and so we are unable to operate the fast up-to-date check. Previously the "state" field would be `null`, and we would return that we weren't up to date. With this change, we don't have a state field, so need to explicitly detect command line mode and bail out when it is enabled. That logic belongs in the VS (host) layer, and so this change introduces a new API via which the host can answer questions that the up-to-date check has for it.

The `LastVersionSeen` value is no longer required, as we now synchronise during the request. This allowed some code to be removed.

Another consequence of the change in Dataflow handling is on the unit tests. Previously we could stage state in the field discussed above. With this change, the field no longer exists. To work around this, an interface was introduced for `Subscription` which allows the logic to be mocked. This change eventually led to much clearer test code.

I've verified that there is no regression in build performance due to this change in the case that the solution has been open long enough for things to stabilise. In a future PR, I will make the fast up-to-date check work immediately after solution open. In such a case, it's important that the up-to-date check waits for the first data to arrive, which the change here achieves. That future PR is the motivation for this change.