Fix hang during up-to-date check
In #7427 we change the fast up-to-date check to make it wait for the latest project data before returning its answer.

This involved a call to `GetLatestVersionAsync`, which is then awaited.

That implementation had a bug that caused a hang while waiting. The up-to-date check is called on the UI thread. It is possible that the dataflow graph, while producing the value we are waiting for, needs access to the UI thread. As we are holding it while waiting, progress cannot be made.

The fix is to "join" the two set of joinable tasks, such that the UI thread is shared between them.