Enable Hot Reload during debugging
Currently, we only enable Hot Reload in non-debugging (i.e., Ctrl+F5) scenarios, and equivalent functionality is provided by Edit and Continue (EnC) in debugging scenarios. Under the hood, however, EnC and Hot Reload are use mostly the same code paths and are really just two different names for the same feature.

And now that the feature is orthogonal to debugging, we would like to have one way to turn it on that is not tied to whether or not we are debugging the running process. Here we remove the check in `ProjectLaunchTargetsProvider` that limited Hot Reload to Ctrl+F5 scenarios.

However, there are some aspects of Hot Reload that need to understand whether or not the process is running under the debugger, and we are now also responsible for passing that information along. This is largely a matter of calling a different overload of `IHotReloadAgentManagerClient.AgentStartedAsync`, which in turn requires a number of changes to pipe the value through.