Opt into Load context binds for package
https://github.com/dotnet/project-system/pull/7447 opt'd Managed.VS into NGEN image.

Before this change, this dll used to be attempted to be loaded in two ways;

- Load context via the MEF cache
- LoadFrom via the [PackageRegistration] and [ProjectObject]

The latter would actually just respect the existing binary in the Load context resulting in a single load.

After NGEN, however, this changed - a Fusion limitation caused to actually load the IL image in the second step, resulting in two versions of the same assembly loaded.

This change forces the second step to use the Load context.