Kill Hot Reload processes when requested
There are going to be times when Hot Reload is in use and the process(es) need to be restarted. Maybe the user has made a code change that was applied to the process, but they don't have a good way of invoking the modified code. In that case we may provide a "Restart Hot Reload" button. Alternatively, they may make a "rude edit"--a code change that would compile, but is outside of the ability of Hot Reload to apply to a running process. In that case we show a dialog offering them the option of restarting. Either way, we need to shut down all the processes associated with a Hot Reload session and then execute a Ctrl+F5 to build and start the processes again. If we don't shut them down first, the build will fail.

The only part of this the project system is responsible for is shutting down the processes when requested. A higher-level component that is aware of all the active Hot Reload sessions and can track the status of the restart operation as a whole is responsible for the rest. Indeed, we already have a `StopProjectAsync` method, it just doesn't actually stop the process yet. Here we add that functionality.