Consolidate code
Consolidate most of the interesting code in `LaunchProfileByIdDataProducer`, `LaunchProfileFromProjectDataProducer` and `LaunchProfileDataProducer` into a new type, `ProjectLaunchProfileHandler`.

The first two types do a lot of similar operations to create entities for launch profiles, suggesting that their functionality should be consolidated into the same type. And the third type only exist because similar functionality is split between the other two.

Also, there were a lot of calls to the project's `ExportProvider` to retrieve `UnconfiguredProject`-level services, and any failure to obtain a service was silently ignored. The new `ProjectLaunchProfileHandler` type also exists within the `UnconfiguredProject` scope, allowing us to import them into the constructor and turning any failure to obtain the other services into a MEF error.

Mostly this change is just moving code around. However, there is a third location where we create new launch profile entities: `LaunchProfileActionBase`. And unfortunately it works differently than the other types. The goal is to eventually consolidate that path in here as well.