Add properties for adding resources/settings files
Contributes to https://github.com/dotnet/project-system/issues/7373

The legacy project properties pages embedded the assembly resources and settings files editors in the "application designer". The new UI takes a property-centric view of the project, and delegates these editors to their own standalone experience.

To help users in this transition, we want a link action in the UI that does two things:

1. Create the file if it does not currently exist
1. Opens the file's editor

CPS PR https://devdiv.visualstudio.com/DevDiv/_git/CPS/pullrequest/344654 adds link action handlers for resources and settings files. This change adds pages and properties to surface this to the user.