Populate categories and properties
When running a query to retrieve data, the query engine tells us which properties to include on the created entity, and also handles calling into the appropriate data producers to create the requested child entities (if any) _and_ hooks them up to them parent.

When running an update query that may create a new entity (such as when adding a launch profile), however, we don't know which properties and sub-entities to create. This is a limitation of the current Project Query API.

Currently when we create an entity for a new launch profile we just populate all the "simple" properties like name, command name, and order. Here we add support for adding all the related entities representing categories and properties, and the simple properties of both. Note this does not include any information about the editor, value, or supported values of a property.

In the absence of information about what was requested in the original property, our goal here is to provide the minimum set of data the UI will need after an update query has finished.