Show the language version
Previously, for language version, we showed a link to an explanation of why the language version could not be modified. #7185 tracks removal of this link, and suggests showing the language version instead.

This change moves the link to the help button, and displays the evaluated language version(s) based on target framework(s) in the UI. These remain read-only values.