Look up child on each iteration
Previously this code was computing a look up of children so that the subsequent lookup would be faster. Overall this approach was O(N) on the number of children.

However that approach is not reliable as changes to the tree will re-write child nodes. That meant that as soon as the loop made an update to the tree, the contents of that lookup could be invalid.

The fix here is to look up the child each time around. This converts to O(N^2), but thankfully N is always a small number here.