Update build queue
The VSEngSS-MicroBuild2019 pool is being phased out in favor of a new pool, VSEngSS-MicroBuild2019-1ES. This should be exactly the same setup, just managed by the 1ES team.

We have until the end of September 2021 to switch over, so if we experience problems with this new pool we can fall back to the old one while they are worked out.