Don't force load of VC projects
When a solution contains both VC and SDK-style projects, the `DotNetCoreProjectCompatibilityDetector` can unwittingly force VS to fully load projects rather than serve them from the cache.

It does this by retrieving the `UnconfiguredProject` from the `IVsHierarchy`, which internally QIs a `IVsBrowseObjectContext` interface, which loads the project.

We avoid that load by performing a capability query on the project instead.