Prevent concurrent access
We have a bug in `HotReloadSessionManager.ActivateSessionAsync` such that it is possible for the process to exit _after_ we've hooked the `Process.Exited` event but before we've started the session. The event handler could end up running on another thread and calling `StopSessionAsync` before the session has started. Then `ActivateSessionAsync` ends up starting a session for the exited process--and because the `Exited` event has already fired we don't get another chance to stop it.

To fix this we need to prevent concurrent access to the `HotReloadSessionManager` state. We already protected the `ImmutableDictionary` of active sessions through the use of `ImmutableInterlocked` but this does nothing to protect the state of the objects _within_ the dictionary. In this change we introduce an `AsyncSemaphore` make use of it everywhere we touch the state.

Also, we were not actually `await`ing the session start, which means that even after introducing the semaphore it might still occur after the `Exited` handler ran. This was to avoid a deadlock that has already been mitigated, so we can re-introduce the `await`.