Stop Hot Reload on project unload
Currently, if the user unloads a project or closes the solution while a Hot Reload session is in progress the session remains, but we now lack any way to interact with it. In some situations this "zombie" session can block changes to later sessions or cause other problems.

The fix here is to end a project's active sessions when the project is unloaded. Note that we don't wait for the call to `StopSessionAsync` to complete; we just need the session to be stopped sometime soon, and we don't want to block the unloading of the project.