Avoid disposing DependenciesSnapshotProvider on the UI thread
Previously we handled the ProjectUnloading event, which fired on the UI thread. As called out in #7635, this accounts for 111ms of UI delay when closing Roslyn.sln.

DependenciesSnapshotProvide is in the UnconfiguredProject scope and will be disposed along with the rest of the project's MEF parts without this special handling. That disposal happens async, on the thread pool.

By removing the ProjectUnloading hook, disposal happens a little later, and off the UI thread.